Extract UIImage test helper to a separate file
